price changed to int64, PostgresOrderRepository.Get() was added

+++ internal/application/order/dto.go
@@ -9,7 +9,7 @@ type OrderItemDTO struct {
 	ID        int64
 	ProductID int64
 	Name      string
-	Price     float64
+	Price     int64
 	Quantity  int64
 }
 
@@ -19,7 +19,7 @@ type OrderDTO struct {
 	Status     string
 	CreatedAt  time.Time
 	Items      []OrderItemDTO
-	Total      float64
+	Total      int64
 }
 
 func ToOrderDTO(order *domain_order.Order) OrderDTO {
@@ -30,7 +30,7 @@ func ToOrderDTO(order *domain_order.Order) OrderDTO {
 			ID:        int64(item.ID()),
 			ProductID: int64(item.ProductID()),
 			Name:      string(item.Name()),
-			Price:     float64(item.Price()),
+			Price:     int64(item.Price()),
 			Quantity:  int64(item.Quantity()),
 		})
 	}
@@ -41,6 +41,6 @@ func ToOrderDTO(order *domain_order.Order) OrderDTO {
 		Status:     string(order.Status()),
 		CreatedAt:  order.CreatedAt(),
 		Items:      items,
-		Total:      float64(order.Total()),
+		Total:      int64(order.Total()),
 	}
 }
